+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2024.12.04
+
+### Added
+
+- Designed kernel implementation for ProjectileCalculator component
+
+### Updated
+
+- Changed design to include some OSU conventions, such as `createNewRep()`.
+- Changed design to include Standard methods.
+
 ## 2024.11.29
 
 ### Added